Fixed issues with MLB and ratings
Found an issues with the Rater object. It will still hardcoding the
length of the features in the split_features() function. After fixing
that the code was able to run. However I then found that I had created
some of the codes wrong, specifically  hp and ap. They were the same as
op which caused all three features to calcluate the same. Fixed the yaml
and am running again.